refactor: Split help
